… QRCoderTests/QRGeneratorTests.cs