# System

```
BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.4291/22H2/2022Update)
13th Gen Intel Core i9-13900K, 1 CPU, 32 logical and 24 physical cores
.NET SDK 8.0.202
  [Host]     : .NET 8.0.3 (8.0.324.11423), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.3 (8.0.324.11423), X64 RyuJIT AVX2
```

# Before

| Method           | Mean       | Error    | StdDev   | Gen0     | Gen1    | Allocated  |
|----------------- |-----------:|---------:|---------:|---------:|--------:|-----------:|
| CreateQRCode     |   245.4 us |  4.88 us |  5.80 us |  17.0898 |  0.2441 |  317.99 KB |
| CreateQRCodeLong | 7,531.1 us | 34.13 us | 31.92 us | 265.6250 | 15.6250 | 4920.13 KB |

# After

| Method           | Mean       | Error   | StdDev  | Gen0    | Allocated |
|----------------- |-----------:|--------:|--------:|--------:|----------:|
| CreateQRCode     |   100.7 us | 0.26 us | 0.22 us |  1.2207 |  23.18 KB |
| CreateQRCodeLong | 1,842.1 us | 3.19 us | 2.83 us | 23.4375 | 433.53 KB |

# Notes

Most of the performance improvements in this PR fall into a few categories:
- Use `BitArray` instead of `string`
- Eliminate concatenation and instead write/copy into buffers using offsets and counts
- Preallocation of arrays/lists to proper length
- Eliminate array allocation when unnecessary (e.g. don't call `AddRange`)
- Eliminate LINQ use
- Improve lookup tables / dictionaries

# Progress

So far about half the code has been converted to use `BitArray`.  There is still a notable amount of code that still uses strings that remains to be optimized.  Also the `Polynom` class/constructor is heavily used and should be optimized.