@Shane32 2024-04-30
By the way, I plan to attempt using ArrayPool for further optimizations but not in this PR.