@Shane32 2024-05-01
@codebude I think this PR has enough changes.  The GitHub comparison should be reasonable enough to follow.  I would like to follow up with some other PRs:
- Improve readability
- Fix bug (?) with string handling
- Utilize ArrayPool and other memory optimizations