@Shane32 2024-05-01
The benchmarks are updated in the PR description, showing 60-75% memory reduction and 100-400% speed increase.

Most likely further PRs will further reduce memory requirements, but not increase speed.