@Shane32 2024-05-01
I'm not planning to re-engineer the mathematics, like to use vector math or anything.  While I'm sure it's possible, that's not my forte.