@Shane32 2024-05-03
Eliminating `foreach` across a `BitArray` and implementing `IEquatable` for `Point` additionally cut down memory requirements for small QR codes by 90%, and above 60% for large QR codes.