@Shane32 2024-05-03
While all beneficial changes, the memory usage didn't change due to the arithmetic changes.  The memory usage changes were due to `foreach` across a `BitArray` internally cast a `bool` to an `object` causing a heap allocation.  Ditto for `List<Point>.Contains` due to `Point` not implementing `IEquatable`.  These were accessed across some 'hot' code, and so many unnecessary heap allocations were made.